refactor: move models, viewmodels to packages
