+++ app/src/main/java/com/example/eventapp/models/Admin.java
@@ -1,4 +1,4 @@
-package com.example.eventapp;
+package com.example.eventapp.models;
 
 public class Admin extends User{
     public Admin(String name) {
