+++ app/src/main/java/com/example/eventapp/models/Entrant.java
@@ -1,4 +1,4 @@
-package com.example.eventapp;
+package com.example.eventapp.models;
 
 public class Entrant extends User{
     private String email;
