+++ app/src/main/java/com/example/eventapp/models/Event.java
@@ -1,4 +1,4 @@
-package com.example.eventapp;
+package com.example.eventapp.models;
 
 public class Event {
     private String eventName;
