+++ app/src/main/java/com/example/eventapp/models/Organizer.java
@@ -1,4 +1,4 @@
-package com.example.eventapp;
+package com.example.eventapp.models;
 
 public class Organizer extends User{
     public Organizer(String name) {
