+++ app/src/main/java/com/example/eventapp/ui/events/EventAdapter.java
@@ -7,7 +7,7 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.example.eventapp.Event;
+import com.example.eventapp.models.Event;
 import com.example.eventapp.R;
 
 import java.util.ArrayList;
